@@ -3,6 +3,7 @@ using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using WeatherLink.Models;
 
@@ -19,8 +20,7 @@ namespace WeatherLink
 		/// <param name="args">Program args</param>
         public static void Main(string[] args)
         {
-			var host = CreateWebHostBuilder(args).Build();
-			host.Run();
+			CreateHostBuilder(args).Build().Run();
         }
 
 		/// <summary>
@@ -28,9 +28,12 @@ namespace WeatherLink
 		/// </summary>
 		/// <param name="args">WebHost args</param>
 		/// <returns>WebHost builder</returns>
-		public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-			WebHost.CreateDefaultBuilder(args)
-				.UseApplicationInsights()
-				.UseStartup<Startup>();
+		public static IHostBuilder CreateHostBuilder(string[] args) =>
+			Host.CreateDefaultBuilder(args)
+				//.UseApplicationInsights()
+				.ConfigureWebHostDefaults(webBuilder =>
+				{
+					webBuilder.UseStartup<Startup>();
+				});
 	}
 }
